Use build flags to centrally choose sensors
Most sensors are optional except for the humidity and air temperature.
To make it easier to choose the sensors in use, this PR adds build flags which also have the pin numbers for each sensor and instructions on how to set them. By default, when a sensor is not enabled, the value is `-1` to signify invalid.

Sometimes one might make a mistake and choose the wrong pin for a sensor or not configure a sensor that is required. A new `config.h` file validates the configuration.

Also removed the SPI library because it is inbuilt in the framework.