Move sensor setup and reading logic to separate files
Moving most of the logic for sensors to separate files/module to keep the main.cpp file easy to manage/maintain. With this it should be easier/faster to add support for other sensors without the burden of scrolling up and down the main.cpp file. It could also allow for supporting other OSS frameworks if need be.

The new `FuFarmSensorsData` struct allows for new data to be added without extra method arguments when transmitting.
The new `FuFarmSensors` centralizes the reading of data into one function. It also creates room to handle calibration when the logic is ready to be added from the sister repository.