Sensor calibration via fuse/toggle
Sensor calibration has been done in [rpi-shiled repository](https://github.com/farm-urban/fufarm_rpi_arduino_shield) but requires flashing different sketches/binaries. This adds the logic for calibration here by using a toggle or fuse which is basically a pin that can be high or low. This allows calibration to happen without having to re-upload new code.

For simple testing, all one needs to do is connect the toggle pin (defaults to 12) to GND then press reset. Once calibration is complete, remove the connection and reset.

Calibration mode is exclusive and when in it, no data is collected or transmitted.