Add support for Arduino Leonardo
This adds support for Arduino Leonardo in one repository but cleverly using build flags to get things working. Basically, the biggest differences are the pins allowed to interrupt (Leonardo has fewer) and the presence of networking (Leonardo does not but instead prints JSON strings on Serial).

With this we can soon archive the [rpi_shield repository](https://github.com/farm-urban/fufarm_rpi_arduino_shield) because we can compile for multiple boards and target different boards for upload.

Also imported mqtt-io configuration file and upload script though unused in this repository.