Support for ENS160
ENS160 offers AQI (Air Quality Index), TVOC (Total Volatile Organic Compounds) and EC02 (Equivalent C02) useful to measure the environment where a growing system may be placed. The C02 measurement from this sensor is just the equivalent and thus should not really replace the more expensive CO2 sensor.

This sensor is listed in Sensor Setups for the latest system and easier to find.