Convert to float when reading ADC to avoid int overflow (#23)
When reading the values from ADC, we convert them to millivolts which means multiplication by at least 1,000 and division by a number based on the precision of the ADC peripheral/block. To avoid loosing data/precision, we do multiplication first then the division later. However, when the original value is a 16 bit number, we hit an overflow and this results in negative numbers. To fix this, we convert to float during the first operation. We could also convert to `uint32_t` but `float` is more common in the setup we have and they both use  4 bytes.