Validation of sensors selection in build
By default, the code has all sensors enabled to just be sure they all fit together nicely and without errors and mistakes or collisions. However, there is are many more scenarios where only one sensor may be in use and that needs validation.

PlatformIO does not natively support loading extra environment variables in its tools. It would be cool but needs a special/custom python script which is too much at this time. A simpler solution was just adding a new environment. To avoid building for that scenario exclusively, a second environment is setup with a different sensor.

As a result of this, an error in configuration of `sen0217InterruptHandler` was discovered.

Also, the environment variable `CI` is set so that PlatformIO CLI can discover it.