Simpler logic for SEN0217 interrupts
Following works in https://github.com/farm-urban/fufarm_arduino_sensors/pull/26, it seems it would be easier to keep logic for sensors mostly inside the files for sensors. This now makes use of a static instance of `FuFarmSensors` when handling the interrupt. Consequently, there is no need to pass a pointer to a function even when not in use.