Moved Home Assistant logic to own files and use library
This moves the code for Home Assistant from `main.cpp` to `HomeAssistant.{h,cpp}` which continues the maintainability effort.

In so doing, it created an avenue for simplification by using the `dawidchyrzynski/home-assistant-integration` library which has support for more features (more sensors) and we do not have to build much of the logic needed. It also handles reconnection to the server.

The device/chip mac address is now used as its unique identifier. It gets converted to HEX but it will be in reverse order as compared to a MAC e.g. MAC printed as `80:00:5E:43:CA:48` will be device `48ca435e0080` and topics looking like `homeassistant/sensor/48ca435e0080/light/config` This unique identifier is printed on startup.

This opens up the possibility of having sensors sending data at different times such as when using an RTOS, timer or interrupts. We could also send commands to toggle a fan, pump, or open a valve using `HAButton`.

Also
1. fixed C02 -> CO2 and eC02 -> eCO2
2. Use fixed length integers to avoid ambiguity where multiple overloads are available
3. Simpler logic for mock by forcing `HAVE_WIFI=0`