Using optional values
This is an idea on how to make values optional hence allowing runtime configuration decisions for which sensors have data and which do not.

Not all boards we support use the latest C++ so we need to have a wrapper.

It is very much a work in progress and it does not yet work well with AVR boards.