Add more versioning information
In some previous PRs, we added version to the information sent to Home Assistant. The version in that defaulted to `0.1.0`. We are not yet doing versioned releases, this will always remain fixed. We can, however, have some more information of the git environment.

Inspired by <https://github.com/sblantipodi/platformio_version_increment>, we can use a python script to pull this information and write a `Version.h` file.
I have modified the script to pull the data from `version.txt` and git. To get something like `0.1.0+versioning.3d55dd3.dogfood`. The `dogfood` marker indicates that we built from a non-CI environment but can be overridden e.g. `export CI=1; pio run uno_r4_wifi`

In CI there is more build information compared on the local machine as well, this overrides the default values by setting up GitVersion in the workflow. The version would be something like `1.161.2-pr914.95` for PRs or `2.0.3` for tags, or `1.44.6-ci.7` from the main branch.

Whereas, we still need to build every time because of WiFi credentials and other configs, this will change when we add provisioning with mDNS (HA supports this already). In the meantime, the commit SHA shown will help track back the version show on the dashboard to a specific version of the code.