Add channel to listing WiFi networks and option to skip listing networks
By default, we scan and list all available WiFi networks on every boot, but that can be slow and eats into precious SRAM. In situations where the network parameters are already stable and known, it makes sense to bypass the scan entirely.

Furthermore, in dense environments, such as mesh networks with multiple APs broadcasting the same SSID but on different channels, repeated SSID entries at different RSSI values clutter the results and add little value.

This change adds a new `WIFI_SKIP_LIST_NETWORKS` macro to skip the network scan when desired, and it also includes the channel number in the scan results to disambiguate them.

The WiFi scan results are no better tabulated now for easier reading (Oh my! This is so hard in C). In the tabulation, SSIDs are truncated/padded at 32 characters because that is the most common max length.