Enable mDNS-based discovery of Home Assistant for MQTT
This change introduces mDNS-based service discovery for Home Assistant on ESP32, ensuring that the device can automatically locate the MQTT broker without relying on a hardcoded IP or hostname. By querying for the `_home-assistant._tcp.local` service, the device no longer needs manual configuration when the Raspberry Pi’s IP changes or is renamed. Instead, it uses the mDNS responder to find the Home Assistant instance on the local network and connect to MQTT on the known port.

Home Assistant’s Mosquitto add-on does not advertise its own `_mqtt._tcp` service, but Supervisor-based installations always publish `_home-assistant._tcp`. We leverage that broadcast to discover Home Assistant’s hostname. Because the HTTP port (8123) is advertised, the code explicitly overrides the port with `HOME_ASSISTANT_MQTT_PORT` (typically 1883).

Notes:
- To account for modern networks, the IPv6 address is prioritised when available.
- The device broadcasts a service `_fufarm._tcp.local` that can be used to troubleshoot visibility.
- When `HOME_ASSISTANT_MQTT_HOST` is specified, discovery is not done.
- Should the server change IP (happens in congested networks), changed are picked up every 5 min when rescan happens.
- The `uno_wifi_rev2` board can no longer support these additional features except if some other sensors are removed. This is done in CI to ensure it build with discovery.

If a user wants to verify what mDNS records are available, the following commands can be used
|OS|Task|Command|
|--|--|--|
|macOS|List all services|`dns-sd -B _services._dns-sd._tcp local`|
|macOS|List just fufarm ones|`dns-sd -B _fufarm`|
|macOS|List just home-assistant ones|`dns-sd -B _home-assistant`|
|macOS|Get details for one|`dns-sd -L "Farm Urban <mac>`|
|Linux|List all services|`avahi-browse -at`|
|Linux|List just fufarm ones with details|`avahi-browse -rt _fufarm._tcp`|
|Linux|List just home-assistant ones with details|`avahi-browse -rt _home-assistant._tcp`|